feat(user): set create, update, and delete permissions

+++ routes/web.php
@@ -24,11 +24,11 @@ Route::group([
     Route::view('/dashboard', 'welcome');
 
     Route::get('/users', [\App\Http\Controllers\UserController::class, 'index'])->can(\App\Enums\PermissionEnum::view_users());
-    Route::post('/users', [\App\Http\Controllers\UserController::class, 'store']);
-    Route::get('/users/create', [\App\Http\Controllers\UserController::class, 'create']);
-    Route::get('/users/{user}', [\App\Http\Controllers\UserController::class, 'show']);
-    Route::put('/users/{user}', [\App\Http\Controllers\UserController::class, 'update']);
-    Route::delete('/users/{user}', [\App\Http\Controllers\UserController::class, 'destroy']);
+    Route::post('/users', [\App\Http\Controllers\UserController::class, 'store'])->can(\App\Enums\PermissionEnum::create_user());
+    Route::get('/users/create', [\App\Http\Controllers\UserController::class, 'create'])->can(\App\Enums\PermissionEnum::create_user());
+    Route::get('/users/{user}', [\App\Http\Controllers\UserController::class, 'show'])->can(\App\Enums\PermissionEnum::update_user());
+    Route::put('/users/{user}', [\App\Http\Controllers\UserController::class, 'update'])->can(\App\Enums\PermissionEnum::update_user());
+    Route::delete('/users/{user}', [\App\Http\Controllers\UserController::class, 'destroy'])->can(\App\Enums\PermissionEnum::delete_user());
 
     Route::get('/roles', [\App\Http\Controllers\RoleController::class, 'index'])->can(\App\Enums\PermissionEnum::view_roles());
     Route::get('/roles/create', [\App\Http\Controllers\RoleController::class, 'create']);
